change path names in auth
